refactor(card): improving Card component and tests following the review's comments
