--- packages/doc/content/components/card/plan-native.mdx
@@ -1,11 +1,0 @@
-### Reference
-
-Gympass `<Card.Plan />` component can be used to display related gyms plans.
-
-### Usage
-
-<ExpoSnack id="@gympass/yoga-card" theme="light" />
-
-### Props
-
-<PropsTable component="Card.Plan" platform="native" />
