--- packages/doc/content/components/card/plan.mdx
@@ -1,20 +1,0 @@
----
-title: 'Plan'
-metaTitle: 'Card Plan Component'
-metaDescription: 'Plan Cards are used to show Gympass plans.'
----
-
-import PlanWeb from './plan-web.mdx';
-import PlanNative from './plan-native.mdx';
-
-# Card.Plan
-
-<TabbedView>
-  <Tab title="Web">
-    <PlanWeb />
-  </Tab>
-
-  <Tab title="Native">
-    <PlanNative />
-  </Tab>
-</TabbedView>
