+++ packages/yoga/src/Card/index.js
@@ -1,7 +1,6 @@
-import { Card, Footer, Header, Plan } from './web';
+import { Card, Footer, Header } from './web';
 
 Card.Footer = Footer;
 Card.Header = Header;
-Card.Plan = Plan;
 
 export default Card;
