+++ packages/yoga/src/Card/index.native.js
@@ -1,7 +1,6 @@
-import { Card, Footer, Header, Plan } from './native';
+import { Card, Footer, Header } from './native';
 
 Card.Footer = Footer;
 Card.Header = Header;
-Card.Plan = Plan;
 
 export default Card;
