+++ packages/yoga/src/Card/web/Card.jsx
@@ -7,6 +7,7 @@ const CardStyled = styled.div(
       components: { card },
     },
   }) => `
+    position: relative;
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
